Flatten package structure for now
