+++ blob.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/grafana/nanogit/protocol"
 	"github.com/grafana/nanogit/protocol/hash"
-	"github.com/grafana/nanogit/protocol/object"
 )
 
 func (c *clientImpl) GetBlob(ctx context.Context, blobID hash.Hash) ([]byte, error) {
@@ -14,7 +14,7 @@ func (c *clientImpl) GetBlob(ctx context.Context, blobID hash.Hash) ([]byte, err
 		return nil, fmt.Errorf("getting object: %w", err)
 	}
 
-	if obj.Type == object.TypeBlob && obj.Hash.Is(blobID) {
+	if obj.Type == protocol.ObjectTypeBlob && obj.Hash.Is(blobID) {
 		return obj.Data, nil
 	}
